Exho Fixes, DarkRP and Chatcommands
It isn't the cleanest solution, but this one works out of the box with
DarkRP and with Chatcommands.

Details:
Instead of calling OnPlayerChat, we used the concommand say and teamsay
to trigger the chat. This way even when you overwrite GM:OnPlayerChat,
the Chat still works.
Also all of the DarkRP Support Changes don't need to be here anymore.